Move User data model, business logic, routes, plus additional Tournament routes
