+++ aesops/__init__.py
@@ -15,5 +15,7 @@ login.login_view = "login"
 from aesops import routes
 from data_models import model_store
 
+from .blueprints.login_blueprint import login_blueprint
 from .blueprints.tournament_blueprint import tournament_blueprint
+app.register_blueprint(login_blueprint)
 app.register_blueprint(tournament_blueprint)
