+++ aesops/templates/_tournament_header.html
@@ -10,7 +10,7 @@
 <div class="row">
     <div class="col-md-3">
         {% if tournament.current_round == 0 and tournament.allow_self_registration%}
-        <a href="{{ url_for('add_player', tid=tournament.id)}}" class="btn btn-primary">Register</a>
+        <a href="{{ url_for('tournaments.add_player', tid=tournament.id)}}" class="btn btn-primary">Register</a>
         {% endif %}
         {% if tournament.current_round > 0 %}
         <h3>Rounds</h3>
