+++ aesops/templates/player_registration.html
@@ -29,7 +29,7 @@
         {{form.pronouns}}<br>
     </p>
     <input type="submit" value="Register" />
-    <p><a href="{{ url_for('add_player', tid=tournament.id) }}" </p>
+    <p><a href="{{ url_for('tournaments.add_player', tid=tournament.id) }}" </p>
 
 </form>
 {% endblock %}
