+++ aesopstables.py
@@ -2,7 +2,7 @@ from aesops import app, db
 from pairing.match import Match
 from pairing.player import Player
 from data_models.tournaments import Tournament
-from aesops.user import User
+from data_models.users import User
 
 
 @app.shell_context_processor
