+++ data_models/model_store.py
@@ -3,7 +3,7 @@ from aesops import db
 from .tournaments import Tournament
 from pairing.player import Player
 from pairing.match import Match
-from aesops.user import User
+from .users import User
 from top_cut.cut import Cut
 from top_cut.cut_player import CutPlayer
 from top_cut.elim_match import ElimMatch
